+++ src/components/FileManager/hooks/use-tree-additional-buttons.tsx
@@ -4,6 +4,7 @@ import {
   GhostIconButton,
 } from '@epam/ai-dial-ui-kit';
 import { IconCopyMinus } from '@tabler/icons-react';
+import { FILE_MANAGER_ICON_STROKE } from '@/constants/icon';
 import { useMemo, type ReactNode } from 'react';
 
 interface useTreeAdditionalButtonsOptions {
@@ -39,7 +40,12 @@ export const useTreeAdditionalButtons = ({
           disabled={isCollapseAllDisabled}
           size={ElementSize.Small}
           onClick={collapseAll}
-          icon={<IconCopyMinus size={DIAL_ICON_SIZE.LG} stroke={1.5} />}
+          icon={
+            <IconCopyMinus
+              size={DIAL_ICON_SIZE.LG}
+              stroke={FILE_MANAGER_ICON_STROKE}
+            />
+          }
           aria-label="collapse-all"
         />
       </>